Fixes non evacuated energy parameters deps cycles
